refactor(server): extract shared layout formatters
Move duplicated formatLayout and formatPlacedItem helpers from
layout.service.ts and share.service.ts into a shared formatters.ts
module to eliminate duplication.
